fact_checker.py
import boto3
import json
import urllib3

# Initialize AWS clients
sqs = boto3.client('sqs')
dynamodb = boto3.client('dynamodb')

# Define the SQS queue URL and DynamoDB table name
QUEUE_URL = 'https://sqs.<region>.amazonaws.com/<account-id>/<queue-name>'
TABLE_NAME = '<dynamodb-table-name>'

def perform_fact_checking_task(text: str) -> str:
    # Perform the fact-checking task
    # This is a placeholder implementation
    # Replace this with actual fact-checking logic
    result = "accurate" if "true" in text else "false"
    return result

def classify_result(result: str) -> str:
    # Classify the result as accurate, inaccurate, false, or indeterminate
    if result == "accurate":
        return "accurate"
    elif result == "inaccurate":
        return "inaccurate"
    elif result == "false":
        return "false"
    else:
        return "indeterminate"
